ibm_db2 driver for sequalize(dialect db2) i.e. sequelize_db2 is deprecated.
tried to pass connection args for db2 in the cloud from .env using mysql dialect.
does not complain, but it does not connect.  sequalize_db2 is a crap.
Anyhow, for what I intended, sequelize_db2 should be utilized - bad product.

ibm_db2 odbc driver is flaky for windows.
build procedure for win32 is missing solution file in node_modules/ibm_db2/build/binding.sol.
I have downloaded latest msbuild, put it in the path, still does not build properly.

Even in the case of Mac, when it builds, odbc driver is not intended for GraphQL service,
which I tried to implement. Sequelize_db is a bad product.